+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## 0.16.0 - TBD
+### Enhancements
+- Added `-t` and `--tsv` flags to DBN CLI to encode tab-separated values (TSV)
+- Added `delimiter` method to builders for `DynEncoder` and `CsvEncoder` to customize the
+  field delimiter character, allowing DBN to be encoded as tab-separated values (TSV)
 
 ### Breaking changes
 - Changed default for `VersionUpgradePolicy` to `Upgrade`